fix: restore compute_segments(Horz) in autohint + FT_Set_Transform parity
- BUGFIX: restore compute_segments(Dimension::Horz) removed in commit 700dd6bc.
  Without HORZ segments, edge-based phantom advance was skipped, causing
  ~1,500 failures in load_glyph/load_char (advance/x, horiBearingX, metrics).
  After fix: load_glyph 57.3k→58.8k, load_char 2,056→2,106 passes in 60k sample.

- feat: real FT_Set_Transform + load compound parity (was 5 stubbed void → 5/5 real)
- feat: FT_LOAD_IGNORE_TRANSFORM honored in FT_Load_Glyph
- fix: apply_transform() no longer corrupts metrics width/height/bearings
- chore: delete 5 dead *_matrix.json fixture files (rejected by harness)
- feat: 4 new Makefile targets (test-op, test-case, test-list, test-list-ops)
- docs: TESTING_SINGLE_FEATURE.md with runtime estimates and env var reference
- types: AFEdge.fpos i16→i32 for wider transformed coordinate range
- cjk/latin: remove redundant as-i32 casts after fpos type change
- oracle: gen_unified_oracle.c now supports compound set-transform+load

+++ pillow-rs-freetype/src/autohint/types.rs
@@ -276,7 +276,7 @@ pub struct AFSegment {
 #[derive(Debug, Clone, Copy)]
 pub struct AFEdge {
     /// Position in font units (canonical key).
-    pub fpos: i16,
+    pub fpos: i32,
     /// Original scaled position (26.6).
     pub opos: i32,
     /// Current (grid-fitted) position (26.6).
